Source code rearranged to be more go-gettable.
Internal imports fixed. Pg and MySQL imports updated.
