+++ .gitignore
@@ -1,4 +1,5 @@
 shawty
 shawty.exe
 src
+!src/shawty
 pkg
