+++ main.go
@@ -4,9 +4,9 @@ import (
 	"errors"
 	"fmt"
 	log "github.com/3fps/log2go"
-	"github.com/3fps/shawty/data"
-	"github.com/3fps/shawty/utils"
-	"github.com/3fps/shawty/web"
+	"shawty/data"
+	"shawty/utils"
+	"shawty/web"
 	"github.com/gorilla/mux"
 	"net"
 	"net/http"
