+++ src/shawty/utils/safe_base.go
