+++ src/shawty/utils/safe_base_test.go
