+++ src/shawty/web/shawtyjs_controller.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"fmt"
 	log "github.com/3fps/log2go"
-	"github.com/3fps/shawty/data"
-	"github.com/3fps/shawty/utils"
+	"shawty/data"
+	"shawty/utils"
 	"html/template"
 	"net/http"
 	"net/url"
